feature #31: Refactor AMRAP, EMOM, and For Time logic separately
Changed the function updateTimer() in the ViewModel to take care of the emom logic. Now the timer keeps counting down and when the time interval finishes the currentRound updates and the timer keeps counting till the totalWorkoutTime is over.

Update the amrap logic also, so now it keep counting down.