feature: Debug and refactor
Refactoring the code, to make the timer work properly.
The data flow was adjusted to pass right the information between views.
It's not 100% finished, so there are some prints to keep tracking the code.

At the TimerMoel, I changed the TimerDetails, so the information is not ambiguous. It was created an enum with the detail and types.

BlueberryTimerTabView, was changed to receive the inputs via TextField.